Bug 778248: Added new kernel with I2C boot fix and PXE support
The new kernel is based on the same commit as the previous kernel, which is

  https://android.googlesource.com/kernel/omap commit 52f4764

The kernel configuration has been restored as far as possible.

The new kernel fixes an I2C bug, which prevented it to boot from Linaro's
u-boot, and static MAC addresses for the Ethernet adapter. This will allow
PXE to work reliably.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>