Bug 816919: Support lowmemorykiller notifications on PandaBoard
Lowmemorykiller notifications are supposed to improve the cleanup
and re-use of available main memory.

The PandaBoard kernel missed a patch to support notifications from
the lowmemorykiller kernel driver. The patch has been imported into
the source repository, and the kernel binary and extra modules are
being updated with this commit.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>